Move TagPokemon to compute.py. Refactor display.py and tagmon.py into a unified file using the newly written code.
