+++ display.py
@@ -6,16 +6,15 @@ from anki.lang import _
 from aqt import mw
 
 from .utils import *
-from .compute import MultiPokemon
-from .tagmon import TagPokemon
+from .compute import MultiPokemon, TagPokemon
 
 
-def display(istagmon, wholecollection=True):
+def pokemon_display(istagmon, wholecollection=True):
     """
-    Control the generation of the html code to displau.
+    Control the generation of the html code to display.
 
     :param bool istagmon: True to switch to use tag's displau, False for deck.
-    :param bool wholeCollection: True if multiple Pokemon, false if single.
+    :param bool wholecollection: True if multiple Pokemon, false if single.
     :return: The html text to display.
     :rtype: str
     """
@@ -36,37 +35,7 @@ def display(istagmon, wholecollection=True):
     #   else, show Pokemon for either single deck or all subdecks and store in Pokemon
     pokemon = TagPokemon() if istagmon else MultiPokemon(wholecollection)
 
-    # Get the html code to display
     result = _show(pokemon)
-
-    return result
-
-def pokemonDisplay(wholeCollection):
-    """
-    Control the generation of the html code to displau.
-
-    :param bool wholeCollection: True if multiple Pokemon, false if single.
-    :return: The html text to display.
-    :rtype: str
-    """
-
-    # Assign Pokemon Image and Progress Bar folder directory names
-    pkmnimgfolder = currentdirname / "pokemon_images"
-    progressbarfolder = currentdirname / "progress_bars"
-
-    # Move Pokemon Image folder to collection.media folder if not already there
-    # (Anki reads from here when running anki.stats.py)
-    if (not os.path.exists(f"{mediafolder}/pokemon_images")) and os.path.exists(pkmnimgfolder):
-        shutil.copytree(pkmnimgfolder, f"{mediafolder}/pokemon_images")
-    if (not os.path.exists(f"{mediafolder}/progress_bars")) and os.path.exists(progressbarfolder):
-        shutil.copytree(progressbarfolder, f"{mediafolder}/progress_bars")
-
-    # If wholeCollection, get all assigned Pokemon and assign to multideckmon,
-    # else, show Pokemon for either single deck or all subdecks and store into multideckmon/deckmon
-    multideckmon = MultiPokemon(wholeCollection)
-
-    # Get the html code to display
-    result = _show(multideckmon)
 
     return result
 
@@ -90,7 +59,7 @@ def _show(data):
     # If single Pokemon, show centered card
     if type(data) == tuple:
         txt += '<div class="pk-st-single">'
-        txt += card_html(data[0], data[1], data[2], data[3] if len(data) == 4 else "")
+        txt += _card_html(data[0], data[1], data[2], data[3] if len(data) == 4 else "")
     # If multiple Pokemon, show flex row of cards
     elif type(data) == list:
         if len(data) == 1:
@@ -103,7 +72,7 @@ def _show(data):
         _num_pokemon = len(data)
         sortedData = sorted(data, key=lambda k: k[2], reverse=True)
         for pokemon in sortedData:
-            txt += card_html(pokemon[0], pokemon[1], pokemon[2], pokemon[3] if len(pokemon) == 4 else "", multi)
+            txt += _card_html(pokemon[0], pokemon[1], pokemon[2], pokemon[3] if len(pokemon) == 4 else "", multi)
 
     txt += '</div>'
 
@@ -111,12 +80,12 @@ def _show(data):
     return txt
 
 
-def card_html(name, deckid, level, nickname="", multi=False):
+def _card_html(name, source, level, nickname="", multi=False):
     """
     Generate the html text for a Pokemon card.
 
     :param str name: Name of the Pokemon.
-    :param int deckid: Id of the deck the Pokemon belongs to.
+    :param source: Id of the deck or name of the tqg the Pokemon belongs to.
     :param int level: The Pokemon's lvl.
     :param str nickname: Pokemon's nickname, if it has any.
     :param bool multi: True if multiple Pokemon are being rendered.
@@ -134,13 +103,13 @@ def card_html(name, deckid, level, nickname="", multi=False):
     # Name and deck
     card += '<div class="pk-st-card-name">' \
             f'<span><b>{nickname if nickname != "" else name}</b></span>' \
-            f'<span><i>{get_deck_name(deckid)}</i></span>' \
+            f'<span><i>{_get_source_name(source)}</i></span>' \
             '</div>'
     # Level
     card += '<div class="pk-st-card-lvl">' \
             '<span style="text-align: right;">Lvl</span>' \
             '<span style="text-align: right;">' \
-            f'<b>{int(level-50) if in_list("prestige", deckid) else int(level)}</b>' \
+            f'<b>{int(level-50) if _in_list("prestige", source) else int(level)}</b>' \
             '</span>' \
             '</div>' \
             '</div>'
@@ -151,7 +120,7 @@ def card_html(name, deckid, level, nickname="", multi=False):
     #############
     # Image
     #############
-    card += f'<img src="/pokemon_images/{image_name(name, deckid)}.png" class="pk-st-card-img"/>'
+    card += f'<img src="/pokemon_images/{_image_name(name, source)}.png" class="pk-st-card-img"/>'
 
     #############
     # Bottom info
@@ -159,7 +128,7 @@ def card_html(name, deckid, level, nickname="", multi=False):
     card += '<div class="pk-st-card-info" style="margin-top: auto;">' \
             '<div class="pk-st-divider" style="margin-bottom: 10px;"></div>'
     # Held/SP
-    held = held_html(deckid)
+    held = _held_html(source)
     if held != "":
         card += '<div class="pk-st-card-sp">' \
                 '<span><b>SP: </b></span>'
@@ -167,30 +136,36 @@ def card_html(name, deckid, level, nickname="", multi=False):
         card += '</div>'
     # Progress bar
     if name == "Egg":
-        card += f'<span class="pk-st-card-xp">{egg_hatch_text(level)}</span>'
+        card += f'<span class="pk-st-card-xp">{_egg_hatch_text(level)}</span>'
     else:
-        card += f'<img src="/progress_bars/{calculate_xp_progress(level)}.png" class="pk-st-card-xp"/>'
+        card += f'<img src="/progress_bars/{_calculate_xp_progress(level)}.png" class="pk-st-card-xp"/>'
     card += '</div>'
 
     # End card
     card += '</div>'
 
+    # TODO: Add # of Pokemon
+    # Make bottom line using function from stats.py and assign to text_lines
+    # line( text_lines, "<b>Total</b>", "</b>%s Pokémon<b>" % _num_pokemon)
+
     return card
 
 
-def get_deck_name(deckid):
+def _get_source_name(item):
     """
-    Get the name of the deck based on its id.
+    Get the name of the tag or deck based on the input item.
 
-    :param int deckid: Deck's id.
+    :param item: Element to find the source of
     :return: The name of the deck
-    :rtype: str
     """
 
-    return mw.col.decks.name(deckid)
+    if isinstance(item, int):
+        return mw.col.decks.name(item)
+    else:
+        return item
 
 
-def in_list(listname, item):
+def _in_list(listname, item):
     """
     Check if an item is in a list. Mainly used to avoid copy/pasting code
     to open the json files.
@@ -207,12 +182,12 @@ def in_list(listname, item):
     return item in get_json(f"_{listname}list.json", [])
 
 
-def image_name(name, deckid):
+def _image_name(name, source):
     """
     Get the image name based on the Pokemon's name and any special attributes.
 
     :param str name: Pokemon's name.
-    :param int deckid: Id of the deck the Pokemon belongs to.
+    :param source: Id of the deck or tag name the Pokemon belongs to.
     :return: The image name to be used to retrieve it.
     :rtype: str
     """
@@ -220,23 +195,23 @@ def image_name(name, deckid):
     pkmnimgfolder = currentdirname / "pokemon_images"
 
     fullname = name
-    if in_list("everstone", deckid):
+    if _in_list("everstone", source):
         # FIX: name is never declared!u
         if name == "Pikachu":
             fullname += "_Ash" + str(random.randint(1, 5))
-    if in_list("megastone", deckid):
+    if _in_list("megastone", source):
         if any([name + "_Mega" in imgname for imgname in os.listdir(pkmnimgfolder)]):
             fullname += "_Mega"
             if name == "Charizard" or name == "Mewtwo":
                 fullname += config["X_or_Y_mega_evolutions"]
-    if in_list("alolan", deckid):
+    if _in_list("alolan", source):
         if any([name + "_Alolan" in imgname for imgname in os.listdir(pkmnimgfolder)]):
             fullname += "_Alolan"
 
     return fullname
 
 
-def egg_hatch_text(level):
+def _egg_hatch_text(level):
     """
     Get the egg's hatch text.
 
@@ -254,7 +229,7 @@ def egg_hatch_text(level):
         return "Making sounds inside"
 
 
-def calculate_xp_progress(level):
+def _calculate_xp_progress(level):
     """
     Calculate the xp progress for the xp bar based on the given level.
 
@@ -265,11 +240,11 @@ def calculate_xp_progress(level):
     return int(float(20*(float(level) - int(float(level)))))
 
 
-def held_html(deckid):
+def _held_html(source):
     """
     Generate the held html code for the given Pokemon.
 
-    :param int deckId: Id of the deck the Pokemon belongs to.
+    :param source: Id of the deck the Pokemon belongs to.
     :return: The concatenation of all held items' html. Empty if it has no items.
     :rtype: str
     """
@@ -284,72 +259,13 @@ def held_html(deckid):
     megastone_html = '<img src="/pokemon_images/item_Mega_Stone.png" height="25px"/>'
     alolan_html = '<img src="/pokemon_images/item_Alolan_Passport.png" height="25px"/>'
 
-    if in_list("prestige", deckid):
+    if _in_list("prestige", source):
         held += '<span>Prestiged </span>'
-    if in_list("everstone", deckid):
+    if _in_list("everstone", source):
         held += everstone_html
-    if in_list("alolan", deckid):
+    if _in_list("alolan", source):
         held += alolan_html
-    if in_list("megastone", deckid):
+    if _in_list("megastone", source):
         held += megastone_html
 
     return held
-
-
-# Deprecated: For tagmon.py only
-def eggHatchText(level, name):
-    if level < 2:
-        text = ("%s (needs a lot more time to hatch)" % name)
-    elif level < 3:
-        text = ("%s (will take some time to hatch)" % name)
-    elif level < 4:
-        text = ("%s (moves around inside sometimes)" % name)
-    else:
-        text = ("%s (making sounds inside)" % name)
-    return text
-
-
-def pokemonDisplayText(name, id, level, nickname):
-    prestigelist = get_json("_prestigelist.json", [])
-    everstonelist = get_json("_everstonelist.json", [])
-    megastonelist = get_json("_megastonelist.json", [])
-    alolanlist = get_json("_alolanlist.json", [])
-
-    held = ""
-    special = ""
-    everstone_html = '<img src="/pokemon_images/item_Everstone.png" hspace="10">'
-    megastone_html = '<img src="/pokemon_images/item_Mega_Stone.png" hspace="10">'
-    alolan_html = '<img src="/pokemon_images/item_Alolan_Passport.png" hspace="10">'
-    pkmnimgfolder = currentdirname / "pokemon_images"
-
-    level = int(float(level))  # float string such as "1.2"
-
-    displayname = name
-    if nickname:
-        displayname = nickname
-    if name == "Egg":
-        text = eggHatchText(level, displayname)
-    else:
-        if id in prestigelist:
-            text = ("%s (Level %s) - Prestiged" %
-                    (displayname, level - 50))
-        else:
-            text = ("%s (Level %s)" % (displayname, level))
-        if id in everstonelist:
-            held += everstone_html
-            # FIX: name is never declared!
-            if name == "Pikachu":
-                special += "_Ash" + str(random.randint(1, 5))
-        if id in megastonelist:
-            held += megastone_html
-
-            if any([name + "_Mega" in imgname for imgname in os.listdir(pkmnimgfolder)]):
-                special += "_Mega"
-                if name == "Charizard" or name == "Mewtwo":
-                    special += config["X_or_Y_mega_evolutions"]
-        if id in alolanlist:
-            held += alolan_html
-            if any([name + "_Alolan" in imgname for imgname in os.listdir(pkmnimgfolder)]):
-                special += "_Alolan"
-
-    return (text, held, special)
